requirements.txt
# Web Framework
Flask==3.1.2
flask-cors==6.0.2

# Google APIs
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2

# Classeviva Integration
Classeviva.py==1.2.9

# Task Scheduling
APScheduler==3.10.4

# Database
mysql-connector-python==9.1.0

# Environment Configuration
python-dotenv==1.1.1

# Additional Dependencies
requests==2.32.5
